chore: update readme
Additional thoughts that did not suit the readme

About unit tests
I wrote unit tests to cover most of the logic contained in my view models. I also wrote a test for the model mapping. This is verify that my logic is sound and works and help for anyone in the future that will need to maintain, alter or refactor the code.

I did not however add any UI tests. Mainly because I would have to extend my network manager and add ways for the application to stub data. This feels out of scope for this proof of concept but is definitly something I would add in a production environment.

About the layout
The application works in any rotation but I did not put much effort into making a landscape specific design. I started out with a single column table view which works great for portrait. But I would definitly consider doing something more in line with how Netflix display sections instead to improve the layout.